feat(monitoring): Summary KPI·Timeseries 인터벌·지연 SLO 확장 및 문서 정리
- 모니터링 설정: sys_monitoring_configs (V6~V10), LATENCY_SLO_TARGET/CRITICAL_THRESHOLD 코드·시드
- Summary API: kpi.latency에 avgLatency, p50Latency, p99Latency, sloTarget, criticalThreshold, prevAvgLatency 추가
- Summary: 가용성 다운타임 1분 버킷·설정 동적 바인딩, 에러 예산 period·consumedRatio 1.0 cap
- Timeseries: LATENCY_P95 인터벌(1m, 5m, 1h, 1d) 동적 그룹화, P95 percentile_cont, 빈 구간 이전값 채우기, 캐시
- docs: ADMIN_MONITORING_TIMESERIES_INTERVAL_SPEC.md 추가, api-spec README 갱신

+++ dwp-auth-server/src/main/java/com/dwp/services/auth/dto/monitoring/TimeseriesResponse.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class TimeseriesResponse {
     
     private String interval; // "HOUR" | "DAY"
-    private String metric; // "PV" | "UV" | "EVENT" | "API_TOTAL" | "API_ERROR"
-    private List<String> labels; // 시간 라벨 (예: ["2026-01-01", "2026-01-02", ...])
-    private List<Long> values; // 값 배열
+    private String metric; // "PV" | "UV" | "EVENT" | "API_TOTAL" | "API_ERROR" | "RPS" | "API_4XX" | "API_5XX" | "LATENCY_P50" | "LATENCY_P95" | "LATENCY_P99" | "AVAILABILITY"
+    private List<String> labels; // 시간 라벨
+    private List<Double> values; // 값 배열 (count 또는 rate/ms/%)
 }
